Get summary up to spec
