+++ Tests/SwiftPrometheusTests/SwiftPrometheusTests.swift
@@ -15,19 +15,6 @@ final class SwiftPrometheusTests: XCTestCase {
         }
     }
     
-    struct BaseSummaryLabels: SummaryLabels {
-        var quantile: String = ""
-        let myValue: String
-        
-        init() {
-            self.myValue = "*"
-        }
-        
-        init(myValue: String) {
-            self.myValue = myValue
-        }
-    }
-    
     var prom: PrometheusClient!
     
     override func setUp() {
@@ -85,19 +72,5 @@ final class SwiftPrometheusTests: XCTestCase {
         XCTAssertEqual(gaugeTwo.get(), 21)
         
         XCTAssertEqual(gauge.collect(), "# HELP my_gauge Gauge for testing\n# TYPE my_gauge gauge\nmy_gauge 21\nmy_gauge{myValue=\"labels\"} 20")
-    }
-    
-    func testSummary() {
-        let summary = prom.createSummary(forType: Double.self, named: "my_summary", helpText: "Summary for testing", quantiles: [0.5, 0.9, 0.99], labels: BaseSummaryLabels.self)
-        let summaryTwo = prom.createSummary(forType: Double.self, named: "my_summary", helpText: "Summary for testing", quantiles: [0.5, 0.9, 0.99], labels: BaseSummaryLabels.self)
-        
-        summary.observe(1)
-        summary.observe(2)
-        summary.observe(4)
-        summaryTwo.observe(10000)
-        
-        summary.observe(123, .init(myValue: "labels"))
-        
-        XCTAssertEqual(summary.collect(), "# HELP my_summary Summary for testing\n# TYPE my_summary summary\nmy_summary{quantile=\"0.5\", myValue=\"*\"} 4.0\nmy_summary{quantile=\"0.9\", myValue=\"*\"} 10000.0\nmy_summary{quantile=\"0.99\", myValue=\"*\"} 10000.0\nmy_summary_count{myValue=\"*\"} 5.0\nmy_summary_sum{myValue=\"*\"} 10130.0\nmy_summary{quantile=\"0.5\", myValue=\"labels\"} 123.0\nmy_summary{quantile=\"0.9\", myValue=\"labels\"} 123.0\nmy_summary{quantile=\"0.99\", myValue=\"labels\"} 123.0\nmy_summary_count{myValue=\"labels\"} 1.0\nmy_summary_sum{myValue=\"labels\"} 123.0")
     }
 }
